Adding the correct typing in GraphQL schema.

+++ src/server/data/schema.ts
@@ -7,7 +7,7 @@ type Query {
 }
 type Timer {
     title: String,
-    duration: Int
+    duration: Float
 }
 `;
